fix(core): switch /event route to effect and implement queue
